modify create2 test
